requirements.txt
python-telegram-bot==20.7
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
schedule==1.2.0
httpx==0.24.1
anyio==3.7.1
httpcore==0.17.3
sniffio==1.3.0
h11==0.14.0
typing_extensions==4.7.1
